+++ application/controllers/Shopping.php
@@ -16,12 +16,30 @@ class Shopping extends Application {
     }
 
     public function summarize() {
+        // identify all of the order files
+        $this->load->helper('directory');
+        $candidates = directory_map('../data/');
+        $parms = array();
+        foreach ($candidates as $filename) {
+           if (substr($filename,0,5) == 'order') {
+               // restore that order object
+               $order = new Order ('../data/' . $filename);
+            // setup view parameters
+               $parms[] = array(
+                   'number' => $order->number,
+                   'datetime' => $order->datetime,
+                   'total' => $order->total()
+                       );
+            }
+        }
+        $this->data['orders'] = $parms;
         $this->data['pagebody'] = 'summary';
         $this->render('template');  // use the default template
     }
 
     public function keep_shopping() {
-        $stuff = file_get_contents('../data/receipt.md');
+        $order = new Order($this->session->userdata('order'));
+        $stuff = $order->receipt();
         $this->data['receipt'] = $this->parsedown->parse($stuff);
         $this->data['content'] = '';
         
@@ -41,9 +59,9 @@ class Shopping extends Application {
 
     public function neworder() {
         // create a new order if needed
-        if (! $this->session->has_userdata('order')) {
+        if (!$this->session->has_userdata('order')) {
             $order = new Order();
-            $this->session->set_userdata('order',$order);
+            $this->session->set_userdata('order', (array) $order);
         }
 
         $this->keep_shopping();
@@ -61,8 +79,27 @@ class Shopping extends Application {
     public function add($what) {
         $order = new Order($this->session->userdata('order'));
         $order->additem($what);
-        $this->session->set_userdata('order',(array)$order);
+        $this->session->set_userdata('order',(array) $order);
         $this->keep_shopping();
+        redirect('/shopping');
+    }
+    
+    public function checkout() {
+        $order = new Order($this->session->userdata('order'));
+        // ignore invalid requests
+        if (! $order->validate())
+            redirect('/shopping');
+
+        $order->save();
+        $this->session->unset_userdata('order');
+        redirect('/shopping');
+    }
+    
+    public function examine($which) {
+        $order = new Order ('../data/order' . $which . '.xml');
+        $stuff = $order->receipt();
+        $this->data['content'] = $this->parsedown->parse($stuff);
+        $this->render();
     }
 
 }